Bug fix
Changed argument from int to float, added rounding code

+++ T9_Pedal_Bundle.h
@@ -60,7 +60,7 @@ void T9PB_effect03_start(void);
 
 // Effect 4: Delay
 void T9PB_effect04_time(float t);
-void T9PB_effect04_taps(int taps);
+void T9PB_effect04_taps(float taps);
 void T9PB_effect04_start(void);
 void T9PB_effect04_stop(void);
 void effect04_update_params(void);
